@@ -16,7 +16,7 @@
     "test:watch": "npm test -- --watch",
     "coverage": "npm test -- --coverage --no-cache",
     "coverage:watch": "npm run coverage -- --watch",
-    "prepare": "npm run build && npm test && npm run doc",
+    "prepare": "yarn build && yarn test && yarn doc",
     "release": "standard-version && npm pack",
     "doc": "typedoc src/index.ts"
   },
@@ -44,34 +44,37 @@
   "homepage": "https://github.com/svenwiegand/typed-intl/issues",
   "devDependencies": {
     "@types/intl": "^1.2.0",
+    "@types/intl-locales-supported": "1.0.0",
     "@types/intl-messageformat": "^1.3.0",
-    "@types/jest": "^21.1.1",
-    "@types/node": "^8.0.31",
-    "browserify": "^14.4.0",
-    "delete-empty": "^1.0.1",
+    "@types/jest": "^23.3.9",
+    "@types/node": "^10.12.5",
+    "@types/webpack": "4.4.18",
+    "browserify": "^16.2.3",
+    "delete-empty": "^2.0.0",
     "dts-bundle": "^0.7.3",
-    "intl-locales-supported": "^1.0.0",
     "intl": "^1.2.5",
-    "jest": "^20.0.4",
-    "standard-version": "^4.2.0",
-    "ts-jest": "^20.0.10",
-    "ts-loader": "^2.3.3",
-    "tslint": "^5.6.0",
-    "tslint-loader": "^3.5.3",
-    "typedoc": "^0.9.0",
-    "typescript": "2.4.2",
-    "webpack": "3.5.5",
-    "webpack-node-externals": "^1.6.0"
+    "intl-locales-supported": "^1.0.0",
+    "jest": "^23.6.0",
+    "standard-version": "^4.4.0",
+    "ts-jest": "^23.10.4",
+    "ts-loader": "^5.3.0",
+    "ts-node": "7.0.1",
+    "tslint": "^5.11.0",
+    "tslint-loader": "^3.5.4",
+    "typedoc": "^0.13.0",
+    "typescript": "3.1.6",
+    "webpack": "4.25.1",
+    "webpack-cli": "^3.1.2",
+    "webpack-node-externals": "^1.7.2"
   },
   "dependencies": {
-    "core-js": "^2.5.1",
-    "intl-messageformat": "^2.1.0"
+    "core-js": "^2.5.7",
+    "intl-messageformat": "^2.2.0"
   },
   "jest": {
     "transform": {
-      ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+      ".(ts|tsx)": "ts-jest"
     },
-    "mapCoverage": true,
     "collectCoverageFrom": [
       "src/**/*.ts"
     ],
