@@ -1,7 +1,7 @@
-import { LanguageTag } from './LanguageTag';
-import IntlMessageFormat from 'intl-messageformat';
-import NumberFormatOptions = Intl.NumberFormatOptions;
-import DateTimeFormatOptions = Intl.DateTimeFormatOptions;
+import { LanguageTag } from './LanguageTag'
+import IntlMessageFormat from 'intl-messageformat'
+import NumberFormatOptions = Intl.NumberFormatOptions
+import DateTimeFormatOptions = Intl.DateTimeFormatOptions
 
 /**
  * Describes additional formats that can be referenced in formatting strings.
@@ -9,30 +9,30 @@ import DateTimeFormatOptions = Intl.DateTimeFormatOptions;
  * See [Format.JS documentation](https://formatjs.io/guides/message-syntax/) for details.
  */
 export interface FormatOptions {
-    number?: { [key: string]: NumberFormatOptions };
-    date?: { [key: string]: DateTimeFormatOptions };
-    time?: { [key: string]: DateTimeFormatOptions };
+  number?: { [key: string]: NumberFormatOptions }
+  date?: { [key: string]: DateTimeFormatOptions }
+  time?: { [key: string]: DateTimeFormatOptions }
 }
 
 function numberFormat(fractionDigits: number): NumberFormatOptions {
-    return {
-        minimumFractionDigits: fractionDigits,
-        maximumFractionDigits: fractionDigits
-    }
+  return {
+    minimumFractionDigits: fractionDigits,
+    maximumFractionDigits: fractionDigits
+  }
 }
 
 const defaultFormats: FormatOptions = {
-    number: {
-        '#': numberFormat(0),
-        '#.#': numberFormat(1),
-        '#.##': numberFormat(2),
-        '#.###': numberFormat(3),
-        '#.####': numberFormat(4),
-        '#.#####': numberFormat(5),
-    }
-};
+  number: {
+    '#': numberFormat(0),
+    '#.#': numberFormat(1),
+    '#.##': numberFormat(2),
+    '#.###': numberFormat(3),
+    '#.####': numberFormat(4),
+    '#.#####': numberFormat(5)
+  }
+}
 
-let _formats: FormatOptions = defaultFormats;
+let _formats: FormatOptions = defaultFormats
 
 /**
  * Global [[FormatOptions]] used by the [[format]] and [[formatObject]] function if nothing else is
@@ -43,27 +43,27 @@ let _formats: FormatOptions = defaultFormats;
  * fraction digits.
  */
 export function formats(): FormatOptions {
-    return _formats as FormatOptions;
+  return _formats as FormatOptions
 }
 
 /**
  * Replaces the current global [[formats]].
  */
 export function setFormats(formatOptions: FormatOptions): void {
-    _formats = formatOptions;
+  _formats = formatOptions
 }
 
 /**
  * Merges the specified formats into the global [[formats]].
  */
 export function addFormats(formatOptions: FormatOptions): void {
-    for (let key in formatOptions) {
-        if (_formats.hasOwnProperty(key)) {
-            _formats[key] = { ... _formats[key] as object, ... formatOptions[key] as object };
-        } else {
-            _formats[key] = formatOptions[key];
-        }
+  for (let key of ['number', 'time', 'date'] as (keyof FormatOptions)[]) {
+    if (_formats.hasOwnProperty(key)) {
+      _formats[key] = { ...(_formats[key] as object), ...(formatOptions[key] as object) }
+    } else {
+      _formats[key] = formatOptions[key]
     }
+  }
 }
 
 /**
@@ -82,10 +82,12 @@ export function addFormats(formatOptions: FormatOptions): void {
  *  Defaults to the globally registered [[formats]].
  * @returns a function accepting the parameter object.
  */
-export function formatObject<P>(language: LanguageTag,
-                                msgFormat: string,
-                                formatOptions: FormatOptions = formats()): (parameters: P) => string {
-    return new IntlMessageFormat(msgFormat, language.tag, formatOptions).format;
+export function formatObject<P>(
+  language: LanguageTag,
+  msgFormat: string,
+  formatOptions: FormatOptions = formats()
+): (parameters: P) => string {
+  return new IntlMessageFormat(msgFormat, language.tag, formatOptions).format
 }
 
 /**
@@ -105,23 +107,30 @@ export function formatObject<P>(language: LanguageTag,
  * @returns a function accepting the specified number of parameters.
  */
 export function format<P1, P2 = undefined, P3 = undefined, P4 = undefined, P5 = undefined>(
-    language: LanguageTag,
-    msgFormat: string,
-    formatOptions: FormatOptions = formats()): (p1: P1, p2?: P2, p3?: P3, p4?: P4, p5?: P5) => string {
-    return (p1: P1, p2?: P2, p3?: P3, p4?: P4, p5?: P5) =>
-        new IntlMessageFormat(msgFormat, language.tag, formatOptions).format({1: p1, 2: p2, 3: p3, 4: p4, 5: p5});
+  language: LanguageTag,
+  msgFormat: string,
+  formatOptions: FormatOptions = formats()
+): (p1: P1, p2?: P2, p3?: P3, p4?: P4, p5?: P5) => string {
+  return (p1: P1, p2?: P2, p3?: P3, p4?: P4, p5?: P5) =>
+    new IntlMessageFormat(msgFormat, language.tag, formatOptions).format({
+      1: p1,
+      2: p2,
+      3: p3,
+      4: p4,
+      5: p5
+    })
 }
 
 /**
  * Defines the different messages for a [[plural]]ized message.
  */
 export interface Plural {
-    zero?: string;
-    one?: string;
-    two?: string;
-    few?: (n: number) => string;
-    many?: (n: number) => string;
-    other: (n: number) => string;
+  zero?: string
+  one?: string
+  two?: string
+  few?: (n: number) => string
+  many?: (n: number) => string
+  other: (n: number) => string
 }
 
 /**
@@ -144,30 +153,31 @@ export interface Plural {
  * @returns a function accepting a numeral argument used to pick the right message.
  */
 export function plural(language: LanguageTag, p: Plural): (n: number) => string {
-    const msgFormat = '{1, plural, ' +
-        (p.zero ? '=0 {zero} ' : '') +
-        (p.one ? 'one {one} ' : '') +
-        (p.two ? 'two {two} ' : '') +
-        (p.few ? 'few {few} ' : '') +
-        (p.many ? 'many {many} ' : '') +
-        'other {other}}';
-    const selection = format<number>(language, msgFormat);
-    return (n: number) => {
-        switch (selection(n)) {
-            case 'zero':
-                return p.zero!;
-            case 'one':
-                return p.one!;
-            case 'two':
-                return p.two!;
-            case 'few':
-                return p.few!(n);
-            case 'many':
-                return p.many!(n);
-            default:
-                return p.other(n);
-        }
-    };
+  const msgFormat =
+    '{1, plural, ' +
+    (p.zero ? '=0 {zero} ' : '') +
+    (p.one ? 'one {one} ' : '') +
+    (p.two ? 'two {two} ' : '') +
+    (p.few ? 'few {few} ' : '') +
+    (p.many ? 'many {many} ' : '') +
+    'other {other}}'
+  const selection = format<number>(language, msgFormat)
+  return (n: number) => {
+    switch (selection(n)) {
+      case 'zero':
+        return p.zero!
+      case 'one':
+        return p.one!
+      case 'two':
+        return p.two!
+      case 'few':
+        return p.few!(n)
+      case 'many':
+        return p.many!(n)
+      default:
+        return p.other(n)
+    }
+  }
 }
 
 /**
@@ -176,8 +186,8 @@ export function plural(language: LanguageTag, p: Plural): (n: number) => string 
  * The `other` case is chosen if non of the more specific cases are matching.
  */
 export interface SelectOptions {
-    [key: string]: string | undefined;
-    other?: string;
+  [key: string]: string | undefined
+  other?: string
 }
 
 /**
@@ -189,8 +199,11 @@ export interface SelectOptions {
  * @param options the possible cases
  * @returns a function expecting a selection string to pick the correct message from the provided options.
  */
-export function select(language: LanguageTag, options: SelectOptions): (selection: string) => string {
-    return selectObject<string>(language, selection => selection, options);
+export function select(
+  language: LanguageTag,
+  options: SelectOptions
+): (selection: string) => string {
+  return selectObject<string>(language, selection => selection, options)
 }
 
 /**
@@ -215,23 +228,26 @@ export function select(language: LanguageTag, options: SelectOptions): (selectio
  * @returns a function expecting a single parameter object containing the selector and
  *     the values referenced by the messages.
  */
-export function selectObject<P>(language: LanguageTag, selector: (parameters: P) => string,
-                                options: SelectOptions): (parameters: P) => string {
-    const optionsFormat: {[key: string]: (parameters: P) => string} = {};
-    for (const option of Object.keys(options)) {
-        const optionMsg = options[option];
-        if (optionMsg) {
-            optionsFormat[option] = formatObject(language, optionMsg);
-        }
+export function selectObject<P>(
+  language: LanguageTag,
+  selector: (parameters: P) => string,
+  options: SelectOptions
+): (parameters: P) => string {
+  const optionsFormat: { [key: string]: (parameters: P) => string } = {}
+  for (const option of Object.keys(options)) {
+    const optionMsg = options[option]
+    if (optionMsg) {
+      optionsFormat[option] = formatObject(language, optionMsg)
     }
-    return parameters => {
-        const s = selector(parameters);
-        const match = optionsFormat[s];
-        const result = match ? match : optionsFormat.other;
-        if (!result) {
-            throw `Selection "${s}" does not match any of the available options`;
-        } else {
-            return result(parameters);
-        }
+  }
+  return parameters => {
+    const s = selector(parameters)
+    const match = optionsFormat[s]
+    const result = match ? match : optionsFormat.other
+    if (!result) {
+      throw `Selection "${s}" does not match any of the available options`
+    } else {
+      return result(parameters)
     }
+  }
 }
